--- templates/index.html
@@ -1,32 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>AI Marketing Agent</title>
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet">
-    <style>
-        .container { max-width: 700px; margin-top: 40px; }
-        .form-label { font-weight: 500; }
-    </style>
-</head>
-<body>
-<div class="container">
-    <div class="card p-4">
-        <h2 class="mb-4">AI Marketing Agent</h2>
-        <form method="post" action="/">
-            <div class="mb-3">
-                <label for="product" class="form-label">Product / App / Service</label>
-                <input type="text" class="form-control" id="product" name="product" placeholder="Enter your product, app, or service" required>
-            </div>
-            <div id="dynamic-questions"></div>
-            <div class="mb-3">
-                <label for="user_images" class="form-label">Image file paths (comma separated, optional)</label>
-                <input type="text" class="form-control" id="user_images" name="user_images" placeholder="e.g. img1.jpg, img2.png">
-            </div>
-            <button type="submit" class="btn btn-primary">Generate Ad Options</button>
-        </form>
-    </div>
-</div>
-</body>
-</html>
